mpegmap/HTMLSourcingInbandTracks-3.3-1.html

<!doctype html>
<html>
  <head>
    <title>TC 5.1.4-11 EISS DataCues</title>
    <link rel="author" title="CableLabs" href="html5@cablelabs.com">
    <meta name="flags" content="[requirement flags]">
    <meta name="assert" content="MPEG-2 SPTS, EISS - UA creates DataCues, attributes set.">
    <meta name="timeout" content="long">
    <script src="testmedia.js"></script>
    <!-- Source HTTP server local copy of World Wide Web Consortium (W3C) Test Harness. --> 
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <!-- Script Version: BETA -->
    <!-- About the Media Resource

         To verify DOM elements the details of the test media must be known prior to test execution.
         This test script is bound to the test media file sourced in the script.
         Sourcing a differnt test media file, or modifying that file, may invalidate the test.

         The media is an MPEG-2, Single Program Transport Stream
         containing two EISS streams, one program insertion cue stream.
    -->
  </head>
  <body>
    <p>
      <a href="http://www.w3.org/TR/html5/references.html#refsINBANDTRACKS">Spec
      Reference [INBANDTRACKS]</a>
    </p>

    <video id="v" width="400" height="300" autoplay controls></video>
    <div id="log"></div>
    <script>
      var test = async_test("EISS DataCues attributes.");
      var vid = document.getElementById("v");
      vid.src = mediaServerURL() + "UserPrivateStreams.ts";

      var hexEiss = '0xA2';
      var eissDesc = parseInt(hexEiss, 16);  // EISS Descriptor 0xA2
      var eissPid = 1768;  // EISS PID

      var hexEbif = '0xA1';
      var ebifDesc = parseInt(hexEbif, 16);  // EBIF Descriptor 0xA1
      var ebifPid = 1770;  // EBIF PID
      var indexEiss = -1;
      var indexEbif = -1;

      // TextTrack[index].cues is a list.  Each cue in the list will have a unique data attribute value.
      // This script tests that all cues are instanceof ArrayBuffer.
      // This script tests TextTrack[index].cues[0].data only for value.
      var dataEiss = [224, 0, 114, 0, 0, 0, 3, 0, 0, 8, 0, 255, 255, 255, 0, 1, 0, 224, 94, 1, 1, 0, 0, 0, 0, 0, 0, 0, 100, 16, 82, 0, 80, 108, 105, 100, 58, 47, 47, 105, 98, 46, 116, 118, 119, 111, 114, 107, 115, 46, 99, 111, 109, 47, 67, 97, 98, 108, 101, 108, 97, 98, 115, 95, 78, 97, 116, 105, 111, 110, 97, 108, 95, 101, 116, 118, 95, 115, 116, 114, 101, 97, 109, 95, 99, 111, 110, 102, 105, 103, 47, 109, 97, 105, 110, 97, 112, 112, 47, 49, 46, 48, 47, 109, 97, 105, 110, 95, 112, 114, 46, 112, 114, 90, 3, 153, 38]; 

      var dataEbif = [227 , 64 , 136 , 251 , 251 , 0 , 59 , 176 , 126 , 0 , 1 , 193 , 0 , 0 , 17 , 3 , 16 , 2 , 128 , 0 , 0 , 1 , 255 , 0 , 0 , 105 , 0 , 0 , 0 , 1 , 3 , 216 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 1 , 0 , 1 , 0 , 0 , 2 , 12 , 1 , 60 , 59 , 67 , 97 , 98 , 108 , 101 , 108 , 97 , 98 , 115 , 95 , 78 , 97 , 116 , 105 , 111 , 110 , 97 , 108 , 95 , 101 , 116 , 118 , 95 , 115 , 116 , 114 , 101 , 97 , 109 , 95 , 99 , 111 , 110 , 102 , 105 , 103 , 47 , 109 , 97 , 105 , 110 , 97 , 112 , 112 , 47 , 49 , 46 , 48 , 47 , 109 , 97 , 105 , 110 , 95 , 112 , 114 , 46 , 112 , 114 , 0 , 15 , 14 , 105 , 98 , 46 , 116 , 118 , 119 , 111 , 114 , 107 , 115 , 46 , 99 , 111 , 109 , 225 , 54 , 136 , 221 , 188 , 252 , 142 , 137];

      vid.addEventListener("playing", function() {
        setTimeout(function() {
          for (var i = 0; i < vid.textTracks.length; i++) {
            if (vid.textTracks[i].id == eissPid) {
              vid.textTracks[i].mode = 'hidden';
              indexEiss = i;
            }
            if (vid.textTracks[i].id == ebifPid) {
              vid.textTracks[i].mode = 'hidden';
              indexEbif = i;
            }
          }
        }, 6000);

        setTimeout(function() {
          vid.pause();

          if (indexEiss == -1) {
            test.step(function() {
              assert_unreached("Did not find EBIF TextTrack.id (" + eissPid + ") in DOM.");
            }, "Did not find EBIF TextTrack.id (" + eissPid + ") in DOM.");
            test.done();
          }
          else if (indexEbif == -1) {
            test.step(function() {
              assert_unreached("Did not find EISS TextTrack.id (" + ebifPid + ") in DOM.");
            }, "Did not find EISS TextTrack.id (" + ebifPid + ") in DOM.");
            test.done();

          } else {
            test.step(function() {
              // EISS
              for (var i = 0; i < vid.textTracks[indexEiss].cues.length; i++) {
                var startTime = vid.textTracks[indexEiss].cues[i].startTime;
                var endTime = vid.textTracks[indexEiss].cues[i].endTime;
                var id = vid.textTracks[indexEiss].cues[i].id;

                assert_equals(id, eissDesc, "DataCue.id");

                if (i <= 13) {
                  assert_equals(startTime, 0, "DataCue.startTime");
                } else if (i <= 14) {
                   assert_approx_equals(startTime, 1.65, 0.1, "DataCue.startTime");
                } else if (i <= 15) {
                   assert_approx_equals(startTime, 3.52, 0.1, "DataCue.startTime");
                } else if (i <= 16) {
                   assert_approx_equals(startTime, 4.79, 0.1, "DataCue.startTime");
                } else if (i <= 17) {
                   assert_approx_equals(startTime, 5.59, 0.1, "DataCue.startTime");
                } else if (i <= 18) {
                   assert_approx_equals(startTime, 8.09, 0.1, "DataCue.startTime");
                } else if (i <= 19) {
                   assert_approx_equals(startTime, 10.46, 0.1, "DataCue.startTime");
                }

                assert_greater_than_equal(endTime + 0.01, startTime, "DataCue.endTime");
                assert_true(vid.textTracks[indexEiss].cues[i].data instanceof ArrayBuffer,
                            "DataCues[i] is ArrayBuffer");

                if (i === 0) {
                  var cueData = new Uint8Array(vid.textTracks[indexEiss].cues[0].data);
                  assert_array_equals(dataEiss, cueData, "DataCues[first] data");
                }

                assert_equals(vid.textTracks[indexEiss].cues[i].pauseOnExit, false, "DataCue.pauseOnExit");
                assert_equals(vid.textTracks[indexEiss].cues[i].text, null, "DataCue.text");
              }

              // EBIF
              for (var i = 0; i < vid.textTracks[indexEbif].cues.length; i++) {
                var startTime = vid.textTracks[indexEbif].cues[i].startTime;
                var endTime = vid.textTracks[indexEbif].cues[i].endTime;
                var id = vid.textTracks[indexEbif].cues[i].id;

                assert_equals(id, ebifDesc, "DataCue.id");

                if (i <= 14) {
                  assert_equals(startTime, 0, "DataCue.startTime");
                } else if (i <= 15) {
                   assert_approx_equals(startTime, 1.65, 0.1, "DataCue.startTime");
                } else if (i <= 16) {
                   assert_approx_equals(startTime, 3.15, 0.1, "DataCue.startTime");
                } else if (i <= 17) {
                   assert_approx_equals(startTime, 4.62, 0.1, "DataCue.startTime");
                } else if (i <= 18) {
                   assert_approx_equals(startTime, 5.52, 0.1, "DataCue.startTime");
                } else if (i <= 19) {
                   assert_approx_equals(startTime, 8.03, 0.1, "DataCue.startTime");
                } else if (i <= 20) {
                   assert_approx_equals(startTime, 10.06, 0.1, "DataCue.startTime");
                }

                assert_greater_than_equal(endTime + 0.01, startTime, "DataCue.endTime");
                assert_true(vid.textTracks[indexEbif].cues[i].data instanceof ArrayBuffer,
                            "DataCues[i] is ArrayBuffer");

                if (i === 0) {
                  var cueData = new Uint8Array(vid.textTracks[indexEbif].cues[0].data);
                  assert_array_equals(dataEbif, cueData, "DataCues[second] data");
                }

                assert_equals(vid.textTracks[indexEbif].cues[i].pauseOnExit, false, "DataCue.pauseOnExit");
                assert_equals(vid.textTracks[indexEbif].cues[i].text, null, "DataCue.text");
              }
            }, "EBIF/EISS DataCues attributes.");
            test.done();
          }
        }, 12000);
      }, false);
    </script>
  </body>
</html>
